2/23/2020 7:23PM - Attack
Attacking has been implemented. The enemy will attack you if he gets too close however, if you are in range, you can attack him by clicking on him. The attack range is 128pixels. A black dot will appear when the enemy is < 10 health. The default enemy hp is 10. 2 damage is dealt per attack on the enemy.